pull most stuff into songparserutil
