+++ game/songparserutil.hh
@@ -1,5 +1,8 @@
 #pragma once
 
+#include "song.hh"
+
+#include <istream>
 #include <string>
 
 namespace SongParserUtil {
@@ -17,6 +20,21 @@ namespace SongParserUtil {
 	void assign(bool& var, std::string const& str);
 	/// Erase last character if it matches
 	void eraseLast(std::string& s, char ch = ' ');
+
+	/// Read a line, tracking the line number for error reporting
+	bool getLine(std::istream& stream, std::string& line, unsigned& linenum);
+	/// Find the BPM in effect at the given timestamp
+	Song::BPM getBPM(Song const& song, double ts);
+	/// Convert a timestamp (beats) into time (seconds); gap is the fallback offset used before any BPM is defined
+	double tsTime(Song const& song, double ts, double gap);
+	/// Append a BPM change at the given timestamp
+	void addBPM(Song& song, double ts, float bpm, double gap);
+	/// Auto-detect cover/background/video/audio files that weren't specified explicitly
+	void guessFiles(Song& song);
+	/// Merge per-singer vocal tracks into a combined "Together" track
+	void vocalsTogether(Song& song);
+	/// Post-process notes after parsing: merge duet tracks, fix overlaps/empty sentences, compute score, add beat markers
+	void finalize(Song& song, unsigned tsPerBeat, unsigned tsEnd, double gap);
 }
